requirements.txt
redis==4.1.3
flower==1.2.0
django==3.2
djangorestframework==3.12.4
requests==2.27.1
django-redis==5.2.0
celery==5.1.2
psycopg2==2.9.5
selenium==3.8.0
beautifulsoup4==4.11.2
lxml



